ref: typehints + imports
